+++ src/main/java/de/bild/codec/SetTypeCodec.java
@@ -1,7 +1,6 @@
 package de.bild.codec;
 
 import java.lang.reflect.Constructor;
-import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -27,25 +26,5 @@ public class SetTypeCodec<C extends Set<V>, V> extends CollectionTypeCodec<C, V>
             return super.getDefaultConstructor((Class) LinkedHashSet.class);
         }
         return super.getDefaultConstructor(clazz);
-    }
-
-
-    /**
-     * Tries to find the implemented interface List with the correct argument and if found returns the correct codec
-     * @param type the type to be examined
-     * @param typeCodecRegistry codec registry for any type
-     * @return returns a SetTypeCodec if type is a set type or null otherwise
-     */
-    public static SetTypeCodec getCodecIfApplicable(Type type, TypeCodecRegistry typeCodecRegistry) {
-        Class rawClass = ReflectionHelper.extractRawClass(type);
-
-        if (rawClass != null && Set.class.isAssignableFrom(rawClass)) {
-            Type setInterface = ReflectionHelper.findInterface(type, Set.class);
-            if (setInterface instanceof ParameterizedType) {
-                ParameterizedType parameterizedType = (ParameterizedType) setInterface;
-                return new SetTypeCodec(rawClass, parameterizedType.getActualTypeArguments()[0], typeCodecRegistry);
-            }
-        }
-        return null;
     }
 }